Troubleshooting Navmesh
Created Probuilder NavMesh object, as an augmentation of Area Target. This creates a 120 vertice NavMesh instead of a 2000 vertice NavMesh if using NavMesh prefab from Vuforia scan.
Seems to work better overall.

NavMeshSurface has inspector property to chose layer to use as navmesh, make sure this is set properly, otherwise its set to “All Layers” by default.

Testing: When phone is locked halfway, and then resumed, the Probuilder Navmesh object still aligns with the Area Target. However, the NavMesh generated at runtime is way off or disappears somewhere. This is a bug i am not sure how to fix. There is no logical explanation for this.

This may remain a bug for a while. For now, I will ignore.